Certificate generator/requirements.txt
easygui
pillow
pandas
openpyxl==3.1.0